rename AbstractAlgorithm -> GeneralAlgorithm
